more robust check_submodules.sh

+++ check_submodule.sh
@@ -19,8 +19,10 @@ git checkout master
 #git submodule update --recursive --remote
 
 git -C COVID-19 checkout master
-git -C COVID-19 rm --cached -r .
-git -C COVID-19 reset --hard
+if [[ `git -C COVID-19 status --porcelain` ]]; then
+  git -C COVID-19 rm --cached -r .
+  git -C COVID-19 reset --hard
+fi
 git -C COVID-19 pull
 
 if [ -f COVID-19/dati-regioni/dpc-covid19-ita-regioni-$d.csv ]; then
